Refactor: Update Utils include paths and replace string-based immediate values with integers in ASM nodes
