+++ include/nanocc/Transforms/ConstantFolding.hpp
@@ -1,9 +1,5 @@
 #include "nanocc/IR/IR.hpp"
-#include "Base.hpp"
 
-
-
-class ConstantFolding : public Pass {
-public:
-    bool run(IRProgramNode& IRProgram) override;
-};
+namespace nanocc {
+    bool ConstantFoldInstructions(IRProgramNode& IRProgram);
+} // namespace nanocc
