+++ include/nanocc/Utils/Utils.hpp
