+++ lib/Sema/SemaDecl.cpp
@@ -2,7 +2,7 @@
 
 #include "nanocc/AST/AST.hpp"
 #include "nanocc/Sema/Sema.hpp"
-#include "nanocc/Utils.hpp"
+#include "nanocc/Utils/Utils.hpp"
 
 #include "SemaHelper.hpp"
